Added custom sorting, minor changes

+++ backend/src/main/java/lt/scoutress/StatisticsApp/entity/Employee.java
@@ -16,7 +16,7 @@ public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "employee_id")
-    private Long employeeId;
+    private Integer employeeId;
 
     @Column(name = "username")
     private String username;
@@ -48,11 +48,11 @@ public class Employee {
         this.joinDate = LocalDate.now();
     }
 
-    public Long getEmployeeId() {
+    public Integer getEmployeeId() {
         return employeeId;
     }
 
-    public void setEmployeeId(Long employeeId) {
+    public void setEmployeeId(Integer employeeId) {
         this.employeeId = employeeId;
     }
